dionaea - Split connection.c to increase readability
