+++ src/Makefile.am
@@ -23,10 +23,15 @@ dionaea_SOURCES += refcount.c
 dionaea_SOURCES += node_info.c
 dionaea_SOURCES += util.c
 dionaea_SOURCES += connection.c
+dionaea_SOURCES += connection_dtls.c
+dionaea_SOURCES += connection_tcp.c
+dionaea_SOURCES += connection_tls.c
+dionaea_SOURCES += connection_udp.c
 dionaea_SOURCES += modules.c
 dionaea_SOURCES += pchild.c
 dionaea_SOURCES += log.c
 dionaea_SOURCES += signals.c
+dionaea_SOURCES += ssl.c
 dionaea_SOURCES += incident.c
 dionaea_SOURCES += threads.c
 dionaea_SOURCES += bistream.c
