+++ src/connection.c
@@ -74,22 +74,6 @@
 #include "incident.h"
 #include "processor.h"
 
-#define CONOFF(x)							((void *)x - sizeof(struct connection))
-#define CONOFF_IO_IN(x)  					((struct connection *)(((void *)x) - offsetof (struct connection, events.io_in)))
-#define CONOFF_IO_OUT(x) 					((struct connection *)(((void *)x) - offsetof (struct connection, events.io_out)))
-#define CONOFF_LISTEN_TIMEOUT(x) 			((struct connection *)(((void *)x) - offsetof (struct connection, events.listen_timeout)))
-#define CONOFF_CONNECTING_TIMEOUT(x) 		((struct connection *)(((void *)x) - offsetof (struct connection, events.connecting_timeout)))
-#define CONOFF_SUSTAIN_TIMEOUT(x)			((struct connection *)(((void *)x) - offsetof (struct connection, events.sustain_timeout)))
-#define CONOFF_IDLE_TIMEOUT(x) 				((struct connection *)(((void *)x) - offsetof (struct connection, events.idle_timeout)))
-#define CONOFF_DNS_TIMEOUT(x) 				((struct connection *)(((void *)x) - offsetof (struct connection, events.dns_timeout)))
-#define CONOFF_HANDSHAKE_TIMEOUT(x) 		((struct connection *)(((void *)x) - offsetof (struct connection, events.handshake_timeout)))
-#define CONOFF_CLOSE_TIMEOUT(x) 			((struct connection *)(((void *)x) - offsetof (struct connection, events.close_timeout)))
-#define CONOFF_RECONNECT_TIMEOUT(x) 		((struct connection *)(((void *)x) - offsetof (struct connection, events.reconnect_timeout)))
-#define CONOFF_THROTTLE_IO_IN_TIMEOUT(x) 	((struct connection *)(((void *)x) - offsetof (struct connection, events.throttle_io_in_timeout)))
-#define CONOFF_THROTTLE_IO_OUT_TIMEOUT(x) 	((struct connection *)(((void *)x) - offsetof (struct connection, events.throttle_io_out_timeout)))
-#define CONOFF_FREE(x)						((struct connection *)(((void *)x) - offsetof (struct connection, events.free)))
-
-
 
 int ssl_tmp_keys_init(struct connection *con);
 
@@ -1910,2195 +1894,6 @@ void connection_throttle_reset(struct connection_throttle *thr)
 {
 	thr->interval_bytes = 0;
 	thr->sleep_adjust = 0;
-}
-
-
-/*
- *
- * connection tcp
- *
- */
-
-
-
-void connection_tcp_accept_cb (EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	while( 1 )
-	{
-
-		struct sockaddr_storage sa;
-		socklen_t sizeof_sa = sizeof(struct sockaddr_storage);
-
-		int accepted_socket = accept(con->socket, (struct sockaddr *)&sa, &sizeof_sa);
-
-		if( accepted_socket == -1 )
-		{
-			if( errno != EAGAIN && errno != EWOULDBLOCK )
-			{
-				g_warning("accept() failed errno=%i (%s)",  errno, strerror(errno));
-			}
-			break;
-		}
-
-		if( accepted_socket > g_dionaea->limits.fds * 70/100 )
-		{
-			g_warning("Running out of fds, closing connection (fd %i limit %i applied limit %i)",
-					  accepted_socket,
-					  g_dionaea->limits.fds,
-					  g_dionaea->limits.fds * 70/100);
-			close(accepted_socket);
-			continue;
-		}
-
-		struct connection *accepted = connection_new(connection_transport_tcp);
-		connection_set_type(accepted, connection_type_accept);
-		accepted->socket = accepted_socket;
-
-		if( connection_node_set_local(accepted) == false || connection_node_set_remote(accepted) == false )
-		{
-			g_warning("accepting connection failed, closing connection");
-			close(accepted->socket);
-			accepted->socket = -1;
-			// Free connection information but don't report
-			// incident.
-			connection_free_cb(loop, &accepted->events.free, 0, false);
-			continue;
-		}
-
-		g_debug("accept() %i local:'%s' remote:'%s'", accepted->socket, accepted->local.node_string,  accepted->remote.node_string);
-
-		connection_set_nonblocking(accepted);
-
-		accepted->data = con->data;
-
-		// set protocol for accepted connection
-		connection_protocol_set(accepted, &con->protocol);
-
-		// copy connect timeout to new connection
-		ev_timer_init(&accepted->events.idle_timeout, connection_idle_timeout_cb, 0. ,con->events.idle_timeout.repeat);
-
-
-		// create protocol specific data
-		accepted->protocol.ctx = con->protocol.ctx_new(accepted);
-
-		// teach new connection about parent
-		if( con->protocol.origin != NULL )
-			con->protocol.origin(accepted, con);
-
-//		stream_processors_init(accepted);
-		accepted->stats.io_in.throttle.max_bytes_per_second = con->stats.io_in.throttle.max_bytes_per_second;
-		accepted->stats.io_out.throttle.max_bytes_per_second = con->stats.io_out.throttle.max_bytes_per_second;
-		connection_established(accepted);
-
-		struct incident *i;
-		i = incident_new("dionaea.connection.tcp.accept");
-		incident_value_con_set(i, "con", accepted);
-		incident_report(i);
-		incident_free(i);
-
-		i = incident_new("dionaea.connection.link");
-		incident_value_con_set(i, "parent", con);
-		incident_value_con_set(i, "child", accepted);
-		incident_report(i);
-		incident_free(i);
-	}
-
-	if( ev_is_active(&con->events.listen_timeout) )
-	{
-		ev_clear_pending(EV_A_ &con->events.listen_timeout);
-		ev_timer_again(EV_A_  &con->events.listen_timeout);
-	}
-}
-
-void connection_tcp_connecting_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_OUT(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	ev_timer_stop(EV_A_ &con->events.connecting_timeout);
-
-	int socket_error = 0;
-	int error_size = sizeof(socket_error);
-
-
-	int ret = getsockopt(con->socket, SOL_SOCKET, SO_ERROR, &socket_error,(socklen_t *)&error_size);
-
-	if( ret != 0 || socket_error != 0 )
-	{
-		errno = socket_error;
-//		perror("getsockopt");
-//		con->protocol.connect_error(con);
-//    	connection_tcp_disconnect(EV_A_ con);
-
-		ev_io_stop(EV_A_ &con->events.io_out);
-		close(con->socket);
-		con->socket = -1;
-		connection_connect_next_addr(con);
-		return;
-	}
-
-	connection_set_state(con, connection_state_established);
-
-	connection_node_set_local(con);
-	connection_node_set_remote(con);
-
-	g_debug("connection %s -> %s", con->local.node_string, con->remote.node_string);
-
-	connection_established(con);
-}
-
-void connection_tcp_io_in_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	int size, buf_size;
-
-	/* determine how many bytes we can recv */
-#ifdef SIOCINQ
-	if( ioctl(con->socket, SIOCINQ, &buf_size) != 0 )
-		buf_size=1024;
-#else
-	buf_size = 16*1024;
-#endif
-
-	/* always increase by one so we get EOF and data in one callback */
-	buf_size++;
-
-	g_debug("can recv %i bytes", buf_size);
-
-	int recv_throttle = connection_throttle(con, &con->stats.io_in.throttle);
-	int recv_size = MIN(buf_size, recv_throttle);
-
-	g_debug("io_in: throttle can %i want %i", buf_size, recv_size);
-
-	if( recv_size == 0 )
-		return;
-
-	unsigned char buf[buf_size];
-
-	GString *new_in = g_string_sized_new(buf_size);
-
-	while( (size = recv(con->socket, buf, recv_size, 0)) > 0 )
-	{
-		g_string_append_len(new_in, (gchar *)buf, size);
-		recv_size -= size;
-		if( recv_size <= 0 )
-			break;
-	}
-	int lerrno = errno;
-
-	if( con->processor_data != NULL && new_in->len > 0 )
-	{
-		processors_io_in(con, new_in->str, new_in->len);
-	}
-
-	connection_throttle_update(con, &con->stats.io_in.throttle, new_in->len);
-	// append
-	g_string_append_len(con->transport.tcp.io_in, new_in->str, new_in->len);
-
-	if( size==0 )//&& size != MIN(buf_size, recv_throttle) )
-	{
-		g_debug("remote closed connection");
-		if( new_in->len > 0 )
-			con->protocol.io_in(con, con->protocol.ctx, (unsigned char *)con->transport.tcp.io_in->str, con->transport.tcp.io_in->len);
-
-		/*
-		 * the protocol may have disabled the watcher for io_in already
-		 * if so, do not deliver the disconnect
-		 */
-		if( ev_is_active(w) )
-			connection_tcp_disconnect(con);
-	} else
-		if( (size == -1 && lerrno == EAGAIN) ||
-			size == MIN(buf_size, recv_throttle) ||
-			recv_size <= 0 )
-	{
-		g_debug("EAGAIN");
-		if( ev_is_active(&con->events.idle_timeout) )
-			ev_timer_again(EV_A_  &con->events.idle_timeout);
-
-		int consumed = 0;
-
-		if( new_in->len > 0 )
-			consumed = con->protocol.io_in(con, con->protocol.ctx, (unsigned char *)con->transport.tcp.io_in->str, con->transport.tcp.io_in->len);
-
-		g_string_erase(con->transport.tcp.io_in, 0, consumed);
-
-		if( con->transport.tcp.io_out->len > 0 && !ev_is_active(&con->events.io_out) )
-			ev_io_start(EV_A_ &con->events.io_out);
-
-	} else
-	{
-		g_warning("recv failed size %i recv_size %i (%s)", size, recv_size, strerror(lerrno));
-		connection_tcp_disconnect(con);
-	}
-	g_string_free(new_in, TRUE);
-
-	if( connection_stats_accounting_limit_exceeded(&con->stats.io_in) )
-	{
-		g_debug("con %p io_in limit exceeded", con);
-		connection_tcp_disconnect(con);
-	}
-}
-
-void connection_tcp_io_out_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_OUT(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	int send_throttle = connection_throttle(con, &con->stats.io_out.throttle);
-	int send_size = MIN(con->transport.tcp.io_out->len, send_throttle);
-
-
-	if( send_size == 0 )
-		return;
-
-
-	int size = send(con->socket, con->transport.tcp.io_out->str, send_size, 0);
-
-	if( ev_is_active(&con->events.idle_timeout) )
-		ev_timer_again(EV_A_  &con->events.idle_timeout);
-
-	if( size > 0 )
-	{
-		connection_throttle_update(con, &con->stats.io_out.throttle, size);
-
-		if( con->processor_data != NULL && size > 0 )
-		{
-			processors_io_out(con, con->transport.tcp.io_out->str, size);
-		}
-
-//		bistream_data_add(&con->bistream, bistream_out, con->transport.tcp.io_out->str, size);
-		g_string_erase(con->transport.tcp.io_out, 0 , size);
-		if( con->transport.tcp.io_out->len == 0 )
-		{
-			if( ev_is_active(&con->events.io_out) )
-				ev_io_stop(EV_A_ w);
-			if( con->state == connection_state_close )
-				connection_tcp_disconnect(con);
-			else
-				if( con->protocol.io_out != NULL )
-			{ /* avoid recursion at any costs */
-				connection_flag_set(con, connection_busy_sending);
-				con->protocol.io_out(con, con->protocol.ctx);
-				connection_flag_unset(con, connection_busy_sending);
-				if( con->transport.tcp.io_out->len > 0 )
-					ev_io_start(CL, &con->events.io_out);
-			}
-		} else
-		{
-			if( !ev_is_active(&con->events.io_out) )
-				ev_io_start(EV_A_ w);
-		}
-	} else
-		if( size == -1 )
-	{
-		if( errno == EAGAIN || errno == EWOULDBLOCK )
-		{
-			if( !ev_is_active(&con->events.io_out) )
-				ev_io_start(CL, &con->events.io_out);
-		} else
-			if( revents != 0 )
-			connection_tcp_disconnect(con);
-	}
-
-	if( connection_stats_accounting_limit_exceeded(&con->stats.io_out) )
-	{
-		g_debug("con %p io_out limit exceeded", con);
-		connection_tcp_disconnect(con);
-	}
-
-}
-
-void connection_tcp_disconnect(struct connection *con)
-{
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	enum connection_state state = con->state;
-	connection_set_state(con, connection_state_close);
-	connection_disconnect(con);
-
-	g_string_erase(con->transport.tcp.io_in, 0, -1);
-	g_string_erase(con->transport.tcp.io_out, 0, -1);
-
-	if( con->protocol.disconnect != NULL &&
-		(state != connection_state_none &&
-		 state != connection_state_connecting ) )
-	{
-		bool reconnect = con->protocol.disconnect(con, con->protocol.ctx);
-		g_debug("reconnect is %i", reconnect);
-		if( reconnect == true && con->type == connection_type_connect )
-		{
-			connection_reconnect(con);
-			return;
-		}
-	}
-	connection_free(con);
-}
-
-
-/*
- *
- * connection ssl
- *
- */
-
-/*
- * the ssl dh key setup is taken from the mod_ssl package from apache
- */
-
-#ifndef SSLC_VERSION_NUMBER
-#define SSLC_VERSION_NUMBER 0x0000
-#endif
-
-DH *myssl_dh_configure(unsigned char *p, int plen,
-					   unsigned char *g, int glen)
-{
-	DH *dh;
-
-	if( !(dh = DH_new()) )
-	{
-		return NULL;
-	}
-
-#if defined(OPENSSL_VERSION_NUMBER) || (SSLC_VERSION_NUMBER < 0x2000)
-	dh->p = BN_bin2bn(p, plen, NULL);
-	dh->g = BN_bin2bn(g, glen, NULL);
-	if( !(dh->p && dh->g) )
-	{
-		DH_free(dh);
-		return NULL;
-	}
-#else
-	R_EITEMS_add(dh->data, PK_TYPE_DH, PK_DH_P, 0, p, plen, R_EITEMS_PF_COPY);
-	R_EITEMS_add(dh->data, PK_TYPE_DH, PK_DH_G, 0, g, glen, R_EITEMS_PF_COPY);
-#endif
-
-	return dh;
-}
-
-
-
-
-
-/*
- * Handle the Temporary RSA Keys and DH Params
- */
-
-
-/*
-** Diffie-Hellman-Parameters: (512 bit)
-**     prime:
-**         00:9f:db:8b:8a:00:45:44:f0:04:5f:17:37:d0:ba:
-**         2e:0b:27:4c:df:1a:9f:58:82:18:fb:43:53:16:a1:
-**         6e:37:41:71:fd:19:d8:d8:f3:7c:39:bf:86:3f:d6:
-**         0e:3e:30:06:80:a3:03:0c:6e:4c:37:57:d0:8f:70:
-**         e6:aa:87:10:33
-**     generator: 2 (0x2)
-** Diffie-Hellman-Parameters: (1024 bit)
-**     prime:
-**         00:d6:7d:e4:40:cb:bb:dc:19:36:d6:93:d3:4a:fd:
-**         0a:d5:0c:84:d2:39:a4:5f:52:0b:b8:81:74:cb:98:
-**         bc:e9:51:84:9f:91:2e:63:9c:72:fb:13:b4:b4:d7:
-**         17:7e:16:d5:5a:c1:79:ba:42:0b:2a:29:fe:32:4a:
-**         46:7a:63:5e:81:ff:59:01:37:7b:ed:dc:fd:33:16:
-**         8a:46:1a:ad:3b:72:da:e8:86:00:78:04:5b:07:a7:
-**         db:ca:78:74:08:7d:15:10:ea:9f:cc:9d:dd:33:05:
-**         07:dd:62:db:88:ae:aa:74:7d:e0:f4:d6:e2:bd:68:
-**         b0:e7:39:3e:0f:24:21:8e:b3
-**     generator: 2 (0x2)
-*/
-
-static unsigned char dh512_p[] = {
-	0x9F, 0xDB, 0x8B, 0x8A, 0x00, 0x45, 0x44, 0xF0, 0x04, 0x5F, 0x17, 0x37,
-	0xD0, 0xBA, 0x2E, 0x0B, 0x27, 0x4C, 0xDF, 0x1A, 0x9F, 0x58, 0x82, 0x18,
-	0xFB, 0x43, 0x53, 0x16, 0xA1, 0x6E, 0x37, 0x41, 0x71, 0xFD, 0x19, 0xD8,
-	0xD8, 0xF3, 0x7C, 0x39, 0xBF, 0x86, 0x3F, 0xD6, 0x0E, 0x3E, 0x30, 0x06,
-	0x80, 0xA3, 0x03, 0x0C, 0x6E, 0x4C, 0x37, 0x57, 0xD0, 0x8F, 0x70, 0xE6,
-	0xAA, 0x87, 0x10, 0x33,
-};
-static unsigned char dh512_g[] = {
-	0x02,
-};
-
-static DH *get_dh512(void)
-{
-	return myssl_dh_configure(dh512_p, sizeof(dh512_p),
-							  dh512_g, sizeof(dh512_g));
-}
-
-static unsigned char dh1024_p[] = {
-	0xD6, 0x7D, 0xE4, 0x40, 0xCB, 0xBB, 0xDC, 0x19, 0x36, 0xD6, 0x93, 0xD3,
-	0x4A, 0xFD, 0x0A, 0xD5, 0x0C, 0x84, 0xD2, 0x39, 0xA4, 0x5F, 0x52, 0x0B,
-	0xB8, 0x81, 0x74, 0xCB, 0x98, 0xBC, 0xE9, 0x51, 0x84, 0x9F, 0x91, 0x2E,
-	0x63, 0x9C, 0x72, 0xFB, 0x13, 0xB4, 0xB4, 0xD7, 0x17, 0x7E, 0x16, 0xD5,
-	0x5A, 0xC1, 0x79, 0xBA, 0x42, 0x0B, 0x2A, 0x29, 0xFE, 0x32, 0x4A, 0x46,
-	0x7A, 0x63, 0x5E, 0x81, 0xFF, 0x59, 0x01, 0x37, 0x7B, 0xED, 0xDC, 0xFD,
-	0x33, 0x16, 0x8A, 0x46, 0x1A, 0xAD, 0x3B, 0x72, 0xDA, 0xE8, 0x86, 0x00,
-	0x78, 0x04, 0x5B, 0x07, 0xA7, 0xDB, 0xCA, 0x78, 0x74, 0x08, 0x7D, 0x15,
-	0x10, 0xEA, 0x9F, 0xCC, 0x9D, 0xDD, 0x33, 0x05, 0x07, 0xDD, 0x62, 0xDB,
-	0x88, 0xAE, 0xAA, 0x74, 0x7D, 0xE0, 0xF4, 0xD6, 0xE2, 0xBD, 0x68, 0xB0,
-	0xE7, 0x39, 0x3E, 0x0F, 0x24, 0x21, 0x8E, 0xB3,
-};
-static unsigned char dh1024_g[] = {
-	0x02,
-};
-
-static DH *get_dh1024(void)
-{
-	return myssl_dh_configure(dh1024_p, sizeof(dh1024_p),
-							  dh1024_g, sizeof(dh1024_g));
-}
-
-/* ----END GENERATED SECTION---------- */
-
-DH *ssl_dh_GetTmpParam(int nKeyLen)
-{
-	DH *dh;
-
-	if( nKeyLen == 512 )
-		dh = get_dh512();
-	else
-		dh = get_dh1024();
-	return dh;
-}
-
-DH *ssl_dh_GetParamFromFile(char *file)
-{
-	DH *dh = NULL;
-	BIO *bio;
-
-	if( (bio = BIO_new_file(file, "r")) == NULL )
-		return NULL;
-#if 0 //SSL_LIBRARY_VERSION < 0x00904000
-	dh = PEM_read_bio_DHparams(bio, NULL, NULL);
-#else
-	dh = PEM_read_bio_DHparams(bio, NULL, NULL, NULL);
-#endif
-	BIO_free(bio);
-	return(dh);
-}
-
-
-#define MYSSL_TMP_KEY_FREE(con, type, idx) \
-    if (con->transport.tls.pTmpKeys[idx]) { \
-        type##_free((type *)con->transport.tls.pTmpKeys[idx]); \
-        con->transport.tls.pTmpKeys[idx] = NULL; \
-    }
-
-#define MYSSL_TMP_KEYS_FREE(con, type) \
-    MYSSL_TMP_KEY_FREE(con, type, SSL_TMP_KEY_##type##_512); \
-    MYSSL_TMP_KEY_FREE(con, type, SSL_TMP_KEY_##type##_1024)
-
-void ssl_tmp_keys_free(struct connection *con)
-{
-	MYSSL_TMP_KEYS_FREE(con, RSA);
-	MYSSL_TMP_KEYS_FREE(con, DH);
-}
-
-int ssl_tmp_key_init_rsa(struct connection *con, int bits, int idx)
-{
-	if( !(con->transport.tls.pTmpKeys[idx] = RSA_generate_key(bits, RSA_F4, NULL, NULL)) )
-	{
-		g_error("Init: Failed to generate temporary %d bit RSA private key", bits);
-		return -1;
-	}
-
-	return 0;
-}
-
-static int ssl_tmp_key_init_dh(struct connection *con, int bits, int idx)
-{
-	if( !(con->transport.tls.pTmpKeys[idx] = ssl_dh_GetTmpParam(bits)) )
-	{
-		g_error("Init: Failed to generate temporary %d bit DH parameters", bits);
-		return -1;
-	}
-
-	return 0;
-}
-
-#define MYSSL_TMP_KEY_INIT_RSA(s, bits) \
-    ssl_tmp_key_init_rsa(s, bits, SSL_TMP_KEY_RSA_##bits)
-
-#define MYSSL_TMP_KEY_INIT_DH(s, bits) \
-    ssl_tmp_key_init_dh(s, bits, SSL_TMP_KEY_DH_##bits)
-
-int ssl_tmp_keys_init(struct connection *con)
-{
-
-	g_message("Init: Generating temporary RSA private keys (512/1024 bits)");
-
-	if( MYSSL_TMP_KEY_INIT_RSA(con, 512) ||
-		MYSSL_TMP_KEY_INIT_RSA(con, 1024) )
-	{
-		return -1;
-	}
-
-	g_message("Init: Generating temporary DH parameters (512/1024 bits)");
-
-	if( MYSSL_TMP_KEY_INIT_DH(con, 512) ||
-		MYSSL_TMP_KEY_INIT_DH(con, 1024) )
-	{
-		return -1;
-	}
-
-	return 0;
-}
-
-RSA *ssl_callback_TmpRSA(SSL *ssl, int export, int keylen)
-{
-	struct connection *c = (struct connection *)SSL_get_app_data(ssl);
-	int idx;
-
-	g_debug("handing out temporary %d bit RSA key", keylen);
-
-	/* doesn't matter if export flag is on,
-	 * we won't be asked for keylen > 512 in that case.
-	 * if we are asked for a keylen > 1024, it is too expensive
-	 * to generate on the fly.
-	 * XXX: any reason not to generate 2048 bit keys at startup?
-	 */
-
-	switch( keylen )
-	{
-	case 512:
-		idx = SSL_TMP_KEY_RSA_512;
-		break;
-
-	case 1024:
-	default:
-		idx = SSL_TMP_KEY_RSA_1024;
-	}
-
-	return(RSA *)c->transport.tls.pTmpKeys[idx];
-}
-
-/*
- * Hand out the already generated DH parameters...
- */
-DH *ssl_callback_TmpDH(SSL *ssl, int export, int keylen)
-{
-	struct connection *c = (struct connection *)SSL_get_app_data(ssl);
-	int idx;
-
-	g_debug("handing out temporary %d bit DH key", keylen);
-
-	switch( keylen )
-	{
-	case 512:
-		idx = SSL_TMP_KEY_DH_512;
-		break;
-
-	case 1024:
-	default:
-		idx = SSL_TMP_KEY_DH_1024;
-	}
-
-	return(DH *)c->transport.tls.pTmpKeys[idx];
-}
-
-
-
-bool connection_tls_set_certificate(struct connection *con, const char *path, int type)
-{
-	g_debug("%s con %p path %s type %i",__PRETTY_FUNCTION__, con, path, type);
-	int ret = SSL_CTX_use_certificate_file(con->transport.tls.ctx, path, type);
-	if( ret != 1 )
-	{
-		perror("SSL_CTX_use_certificate_file");
-		return false;
-	}
-	return true;
-}
-
-bool connection_tls_set_key(struct connection *con, const char *path, int type)
-{
-	g_debug("%s con %p path %s type %i",__PRETTY_FUNCTION__, con, path, type);
-	int ret = SSL_CTX_use_PrivateKey_file(con->transport.tls.ctx, path, type);
-	if( ret != 1 )
-	{
-		perror("SSL_CTX_use_PrivateKey_file");
-		return false;
-	}
-	return true;
-}
-
-
-int add_ext(X509 *cert, int nid, char *value)
-{
-	X509_EXTENSION *ex;
-	X509V3_CTX ctx;
-	/* This sets the 'context' of the extensions. */
-	/* No configuration database */
-	X509V3_set_ctx_nodb(&ctx);
-	/* Issuer and subject certs: both the target since it is self signed,
-	 * no request and no CRL
-	 */
-	X509V3_set_ctx(&ctx, cert, cert, NULL, NULL, 0);
-	ex = X509V3_EXT_conf_nid(NULL, &ctx, nid, value);
-	if( !ex )
-		return 0;
-
-	X509_add_ext(cert,ex,-1);
-	X509_EXTENSION_free(ex);
-	return 1;
-}
-
-static void callback(int p, int n, void *arg)
-{
-	char c='B';
-
-	if( p == 0 ) c='.';
-	if( p == 1 ) c='+';
-	if( p == 2 ) c='*';
-	if( p == 3 ) c='\n';
-	fputc(c,stderr);
-}
-
-
-bool mkcert(SSL_CTX *ctx)
-{
-	int bits = 512*4;
-	int serial = time(NULL);
-	int days = 365;
-	gchar *value = NULL;
-	GError *error = NULL;
-
-
-	X509 *x;
-	EVP_PKEY *pk;
-	RSA *rsa;
-	X509_NAME *name=NULL;
-
-	if( (pk=EVP_PKEY_new()) == NULL )
-		goto err;
-
-	if( (x=X509_new()) == NULL )
-		goto err;
-
-	rsa=RSA_generate_key(bits,RSA_F4,callback,NULL);
-	if( !EVP_PKEY_assign_RSA(pk,rsa) )
-	{
-		perror("EVP_PKEY_assign_RSA");
-		goto err;
-	}
-	rsa=NULL;
-
-	X509_set_version(x,2);
-	ASN1_INTEGER_set(X509_get_serialNumber(x),serial);
-	X509_gmtime_adj(X509_get_notBefore(x),0);
-	X509_gmtime_adj(X509_get_notAfter(x),(long)60*60*24*days);
-	X509_set_pubkey(x,pk);
-
-	name=X509_get_subject_name(x);
-
-	value = g_key_file_get_string(g_dionaea->config, "dionaea", "ssl.default.c", &error);
-	if (value == NULL) {
-		value = g_strdup("DE");
-	}
-	g_clear_error(&error);
-	X509_NAME_add_entry_by_txt(name, "C", MBSTRING_ASC, (const unsigned char *)value, -1, -1, 0);
-	g_free(value);
-
-	value = g_key_file_get_string(g_dionaea->config, "dionaea", "ssl.default.cn", &error);
-	if (value == NULL) {
-		value = g_strdup("Nepenthes Development Team");
-	}
-	g_clear_error(&error);
-	X509_NAME_add_entry_by_txt(name, "CN", MBSTRING_ASC, (const unsigned char*)value, -1, -1, 0);
-	g_free(value);
-
-	value = g_key_file_get_string(g_dionaea->config, "dionaea", "ssl.default.o", &error);
-	if (value == NULL) {
-		value = g_strdup("dionaea.carnivore.it");
-	}
-	g_clear_error(&error);
-	X509_NAME_add_entry_by_txt(name, "O", MBSTRING_ASC, (const unsigned char*)value, -1, -1, 0);
-	g_free(value);
-
-	value = g_key_file_get_string(g_dionaea->config, "dionaea", "ssl.default.ou", &error);
-	if (value == NULL) {
-		value = g_strdup("anv");
-	}
-	g_clear_error(&error);
-	X509_NAME_add_entry_by_txt(name, "OU", MBSTRING_ASC, (const unsigned char*)value, -1, -1, 0);
-	g_free(value);
-
-
-	/* Its self signed so set the issuer name to be the same as the
-	 * subject.
-	 */
-	X509_set_issuer_name(x,name);
-
-	add_ext(x, NID_netscape_cert_type, "server");
-	add_ext(x, NID_netscape_ssl_server_name, "localhost");
-
-	if( !X509_sign(x,pk,EVP_md5()) )
-		goto err;
-
-
-	int ret = SSL_CTX_use_PrivateKey(ctx, pk);
-	if( ret != 1 )
-	{
-		perror("SSL_CTX_use_PrivateKey");
-		return false;
-	}
-
-	ret = SSL_CTX_use_certificate(ctx, x);
-	if( ret != 1 )
-	{
-		perror("SSL_CTX_use_certificate");
-		return false;
-	}
-
-	return true;
-	err:
-	return false;
-}
-
-bool connection_tls_mkcert(struct connection *con)
-{
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	return mkcert(con->transport.tls.ctx);
-}
-
-void connection_tls_io_out_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = NULL;
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	if( w->events == EV_READ )
-		con = CONOFF_IO_IN(w);
-	else
-		con	= CONOFF_IO_OUT(w);
-
-
-	if( con->transport.tls.io_out_again->len == 0 )
-	{
-		GString *io_out_again = con->transport.tls.io_out_again;
-		con->transport.tls.io_out_again = con->transport.tls.io_out;
-		con->transport.tls.io_out = io_out_again;
-		con->transport.tls.io_out_again_size = 0;
-	}
-
-
-	int send_throttle = connection_throttle(con, &con->stats.io_out.throttle);
-	if( con->transport.tls.io_out_again_size == 0 )
-		con->transport.tls.io_out_again_size = MIN((int)con->transport.tls.io_out_again->len, send_throttle);
-
-	if( con->transport.tls.io_out_again_size <= 0 )
-		return;
-
-	g_debug("send_throttle %i con->transport.tcp.io_out_again->len %i con->transport.ssl.io_out_again_size %i todo %i",
-			send_throttle, (int)con->transport.tls.io_out_again->len, con->transport.tls.io_out_again_size,
-			(int)con->transport.tls.io_out_again->len + (int)con->transport.tls.io_out->len);
-
-
-	int err = SSL_write(con->transport.tls.ssl, con->transport.tls.io_out_again->str, con->transport.tls.io_out_again_size);
-	connection_tls_error(con);
-
-	if( err <= 0 )
-	{
-		int action = SSL_get_error(con->transport.tls.ssl, err);
-		connection_tls_error(con);
-		switch( action )
-		{
-		case SSL_ERROR_ZERO_RETURN:
-			g_debug("%s:%i", __FILE__,  __LINE__);
-			if( revents != 0 )
-				connection_tls_disconnect(con);
-			else
-				connection_set_state(con, connection_state_close);
-			break;
-
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-
-			if( ev_is_active(&con->events.io_in) && revents != EV_READ )
-			{
-				ev_io_stop(CL, &con->events.io_in);
-				ev_io_init(&con->events.io_in, connection_tls_io_out_cb, con->socket, EV_READ);
-				ev_io_start(CL, &con->events.io_in);
-			}
-
-			if( ev_is_active(&con->events.io_out) )
-				ev_io_stop(CL, &con->events.io_out);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			if( !ev_is_active(&con->events.io_out) )
-				ev_io_start(CL, &con->events.io_out);
-
-			if( ev_is_active(&con->events.io_in) )
-				ev_io_stop(CL, &con->events.io_in);
-			break;
-
-		case SSL_ERROR_WANT_ACCEPT:
-			g_debug("SSL_ERROR_WANT_ACCEPT%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_X509_LOOKUP:
-			g_debug("SSL_ERROR_WANT_X509_LOOKUP %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SYSCALL:
-			g_debug("SSL_ERROR_SYSCALL %s:%i", __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-
-		case SSL_ERROR_SSL:
-			g_debug("SSL_ERROR_SSL %s:%i", __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-
-		case SSL_ERROR_NONE:
-			g_debug("%s:%i", __FILE__,  __LINE__);
-			break;
-
-		}
-	} else
-	{
-		int size = err;
-
-		if( size == con->transport.tls.io_out_again_size )
-		{
-			/* restore io handlers to fit default */
-			if( ev_is_active(&con->events.io_in) && ev_cb(&con->events.io_in) != connection_tls_io_in_cb )
-				ev_io_stop(CL, &con->events.io_in);
-
-			if( !ev_is_active(&con->events.io_in) )
-			{
-				ev_io_init(&con->events.io_in, connection_tls_io_in_cb, con->socket, EV_READ);
-				ev_io_start(CL, &con->events.io_in);
-			}
-
-			if( ev_is_active(&con->events.io_out) && ev_cb(&con->events.io_out) != connection_tls_io_out_cb )
-				ev_io_stop(CL, &con->events.io_out);
-
-			if( !ev_is_active(&con->events.io_out) )
-			{
-				ev_io_init(&con->events.io_out, connection_tls_io_out_cb, con->socket, EV_WRITE);
-				ev_io_start(CL, &con->events.io_out);
-			}
-
-			connection_throttle_update(con, &con->stats.io_out.throttle, size);
-
-			g_string_erase(con->transport.tls.io_out_again, 0 , con->transport.tls.io_out_again_size);
-			con->transport.tls.io_out_again_size = 0;
-
-			if( con->transport.tls.io_out_again->len == 0 && con->transport.tls.io_out->len == 0 )
-			{
-				g_debug("connection is flushed");
-				if( ev_is_active(&con->events.io_out) )
-					ev_io_stop(EV_A_ &con->events.io_out);
-
-				if( con->state == connection_state_close )
-					connection_tls_shutdown_cb(EV_A_ w, revents);
-				else
-					if( con->protocol.io_out != NULL )
-				{
-					/* avoid recursion */
-					connection_flag_set(con, connection_busy_sending);
-					con->protocol.io_out(con, con->protocol.ctx);
-					connection_flag_unset(con, connection_busy_sending);
-					if( con->transport.tls.io_out->len > 0 )
-						ev_io_start(CL, &con->events.io_out);
-				}
-			}
-
-
-
-		} else
-		{
-			g_debug("unexpected %s:%i...", __FILE__,  __LINE__);
-		}
-
-	}
-}
-
-
-void connection_tls_shutdown_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = NULL;
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	if( w->events == EV_READ )
-		con = CONOFF_IO_IN(w);
-	else
-		con	= CONOFF_IO_OUT(w);
-
-	if( con->type == connection_type_listen )
-	{
-		g_debug("connection was listening, closing!");
-		connection_tls_disconnect(con);
-		return;
-	}
-
-	if( SSL_get_shutdown(con->transport.tls.ssl) & (SSL_SENT_SHUTDOWN|SSL_RECEIVED_SHUTDOWN) )
-	{
-		g_debug("connection has sent&received shutdown");
-		connection_tls_disconnect(con);
-		return;
-	}
-
-	ev_io_stop(EV_A_ &con->events.io_in);
-	ev_io_stop(EV_A_ &con->events.io_out);
-
-	connection_tls_error(con);
-
-	int err = SSL_shutdown(con->transport.tls.ssl);
-	connection_tls_error(con);
-
-	int action;
-
-	switch( err )
-	{
-	case 1:
-		connection_tls_disconnect(con);
-		break;
-
-	case 0:
-		err = SSL_shutdown(con->transport.tls.ssl);
-		action = SSL_get_error(con->transport.tls.ssl, err);
-		connection_tls_error(con);
-
-		switch( action )
-		{
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-			ev_io_init(&con->events.io_in, connection_tls_shutdown_cb, con->socket, EV_READ);
-			ev_io_start(EV_A_ &con->events.io_in);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			ev_io_init(&con->events.io_out, connection_tls_shutdown_cb, con->socket, EV_WRITE);
-			ev_io_start(EV_A_ &con->events.io_out);
-			break;
-
-		case SSL_ERROR_WANT_ACCEPT:
-			g_debug("SSL_ERROR_WANT_ACCEPT%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_X509_LOOKUP:
-			g_debug("SSL_ERROR_WANT_X509_LOOKUP %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SYSCALL:
-			g_debug("SSL_ERROR_SYSCALL %i %s %s:%i", errno, strerror(errno), __FILE__,  __LINE__);
-			if( errno == 0 )
-			{
-				/*
-				 * HACK actually a bug in openssl - a patch sent on
-				 * 2006-06-29 0:12:51
-				 * with subject
-				 * [PATCH2] Fix for SSL_shutdown() with non-blocking not returning -1
-				 * by Darryl L. Miles
-				 * actually fixes the issue
-				 *
-				 * patch was merged into openssl
-				 * 2009-Apr-07 18:28 http://cvs.openssl.org/chngview?cn=17995
-				 * and will (hopefully) ship with openssl 0.9.8l
-				 *
-				 * given the 3 years it took openssl to accept a patch,
-				 * it did not take me that long to figure out
-				 * why SSL_shutdown failed on nonblocking sockets
-				 *
-				 * at the time of this writing, 0.9.8k is current
-				 * 0.9.8g is shipped by all major vendors as stable
-				 *
-				 * so it may take some time to get this fix to the masses
-				 *
-				 * due to unclear&complex openssl version situation
-				 * I decided not to provide an workaround, just close the connection instead
-				 * and rant about openssl
-				 *
-				 */
-				connection_tls_disconnect(con);
-			}else
-				connection_tls_disconnect(con);
-
-			break;
-
-		case SSL_ERROR_SSL:
-			g_debug("SSL_ERROR_SSL %s:%i", __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-		}
-
-		break;
-
-	case -1:
-		g_debug("SSL_shutdown -1 %s:%i", __FILE__,  __LINE__);
-		action = SSL_get_error(con->transport.tls.ssl, err);
-		connection_tls_error(con);
-
-		switch( action )
-		{
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-			ev_io_init(&con->events.io_in, connection_tls_shutdown_cb, con->socket, EV_READ);
-			ev_io_start(EV_A_ &con->events.io_in);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			ev_io_init(&con->events.io_out, connection_tls_shutdown_cb, con->socket, EV_WRITE);
-			ev_io_start(EV_A_ &con->events.io_out);
-			break;
-
-		default:
-			g_debug("SSL_ERROR_ %i %s:%i", action, __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-		}
-		break;
-
-	default:
-		g_debug("SSL_shutdown %i %s:%i", err, __FILE__,  __LINE__);
-		break;
-	}
-}
-
-void connection_tls_io_in_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = NULL;
-
-	if( w->events == EV_READ )
-		con = CONOFF_IO_IN(w);
-	else
-		con	= CONOFF_IO_OUT(w);
-
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-
-	int recv_throttle = connection_throttle(con, &con->stats.io_in.throttle);
-	if( recv_throttle == 0 )
-	{
-		g_debug("recv throttle %i", recv_throttle);
-		return;
-	}
-
-	unsigned char buf[recv_throttle];
-
-	int err=0;
-	if( (err = SSL_read(con->transport.tls.ssl, buf, recv_throttle)) > 0 )
-	{
-//		g_debug("SSL_read %i %.*s", err, err, buf);
-		g_string_append_len(con->transport.tls.io_in, (gchar *)buf, err);
-	}
-	connection_tls_error(con);
-
-	int action = SSL_get_error(con->transport.tls.ssl, err);
-	connection_tls_error(con);
-
-	if( err<=0 )
-	{
-		switch( action )
-		{
-		case SSL_ERROR_NONE:
-			g_debug("%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_ZERO_RETURN:
-			g_debug("%s:%i", __FILE__,  __LINE__);
-			connection_tls_shutdown_cb(EV_A_ w, revents);
-			break;
-
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-			if( ev_is_active(&con->events.io_out) )
-				ev_io_stop(CL, &con->events.io_out);
-
-			if( !ev_is_active(&con->events.io_in) )
-				ev_io_start(CL, &con->events.io_in);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			if( ev_is_active(&con->events.io_in) )
-				ev_io_stop(CL, &con->events.io_in);
-
-			if( ev_is_active( &con->events.io_out ) && revents != EV_WRITE )
-			{
-				ev_io_stop(EV_A_ &con->events.io_out);
-				ev_io_init(&con->events.io_out, connection_tls_io_in_cb, con->socket, EV_WRITE);
-				ev_io_start(EV_A_ &con->events.io_out);
-			}
-			break;
-
-		case SSL_ERROR_WANT_ACCEPT:
-			g_debug("SSL_ERROR_WANT_ACCEPT%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_X509_LOOKUP:
-			g_debug("SSL_ERROR_WANT_X509_LOOKUP %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SYSCALL:
-			g_debug("SSL_ERROR_SYSCALL %s:%i", __FILE__,  __LINE__);
-			if( err == 0 )
-				g_debug("remote closed protocol, violating the specs!");
-			else
-				if( err == -1 )
-				perror("read");
-
-			connection_tls_disconnect(con);
-			break;
-
-		case SSL_ERROR_SSL:
-			g_debug("SSL_ERROR_SSL %s:%i", __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-		}
-	} else
-		if( err > 0 )
-	{
-
-		/* restore io handlers to fit default */
-		if( ev_is_active(&con->events.io_in) && ev_cb(&con->events.io_in) != connection_tls_io_in_cb )
-			ev_io_stop(CL, &con->events.io_in);
-
-		if( !ev_is_active(&con->events.io_in) )
-		{
-			ev_io_init(&con->events.io_in, connection_tls_io_in_cb, con->socket, EV_READ);
-			ev_io_start(CL, &con->events.io_in);
-		}
-
-		if( ev_is_active(&con->events.io_out) && ev_cb(&con->events.io_out) != connection_tls_io_out_cb )
-			ev_io_stop(CL, &con->events.io_out);
-
-		if( !ev_is_active(&con->events.io_out) )
-		{
-			ev_io_init(&con->events.io_out, connection_tls_io_out_cb, con->socket, EV_WRITE);
-		}
-
-
-		connection_throttle_update(con, &con->stats.io_in.throttle, err);
-		if( ev_is_active(&con->events.idle_timeout) )
-			ev_timer_again(EV_A_  &con->events.idle_timeout);
-
-
-		con->protocol.io_in(con, con->protocol.ctx, (unsigned char *)con->transport.tls.io_in->str, con->transport.tls.io_in->len);
-		con->transport.tls.io_in->len = 0;
-
-		if( (con->transport.tls.io_out->len > 0 || con->transport.tls.io_out_again->len > 0 ) &&
-			!ev_is_active(&con->events.io_out) )
-			ev_io_start(EV_A_ &con->events.io_out);
-	}
-}
-
-void connection_tls_accept_cb (EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	struct incident *i;
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	while( 1 )
-	{
-		struct sockaddr_storage sa;
-		socklen_t sizeof_sa = sizeof(struct sockaddr_storage);
-
-		// clear accept timeout, reset
-
-
-		int accepted_socket = accept(con->socket, (struct sockaddr *)&sa, &sizeof_sa);
-
-		if( accepted_socket == -1 && (errno == EAGAIN || errno == EWOULDBLOCK) )
-			break;
-
-
-		if( accepted_socket > g_dionaea->limits.fds * 70/100 )
-		{
-			g_warning("Running out of fds, closing connection (fd %i limit %i applied limit %i)",
-					  accepted_socket,
-					  g_dionaea->limits.fds,
-					  g_dionaea->limits.fds * 70/100);
-			close(accepted_socket);
-			continue;
-		}
-
-
-		struct connection *accepted = connection_new(connection_transport_tls);
-		SSL_CTX_free(accepted->transport.tls.ctx);
-		connection_set_type(accepted, connection_type_accept);
-		accepted->socket = accepted_socket;
-		accepted->data = con->data;
-
-		connection_node_set_local(accepted);
-		connection_node_set_remote(accepted);
-
-		g_debug("accept() %i local:'%s' remote:'%s'", accepted->socket, accepted->local.node_string,  accepted->remote.node_string);
-		connection_set_nonblocking(accepted);
-
-		// set protocol for accepted connection
-		connection_protocol_set(accepted, &con->protocol);
-
-		accepted->stats.io_out.throttle.max_bytes_per_second = con->stats.io_out.throttle.max_bytes_per_second;
-
-		accepted->transport.tls.ctx = con->transport.tls.ctx;
-		accepted->transport.tls.ssl = SSL_new(accepted->transport.tls.ctx);
-		SSL_set_fd(accepted->transport.tls.ssl, accepted->socket);
-
-		SSL_set_app_data(accepted->transport.tls.ssl, con);
-//		SSL_set_app_data2(ssl, NULL); /* will be request_rec */
-
-//		sslconn->ssl = ssl;
-
-		/*
-		 *  Configure callbacks for SSL connection
-		 */
-//		memcpy(accepted->transport.ssl.pTmpKeys, con->transport.ssl.pTmpKeys, sizeof(void *)*SSL_TMP_KEY_MAX);
-//		accepted->transport.ssl.parent = con;
-		SSL_set_tmp_rsa_callback(accepted->transport.tls.ssl, ssl_callback_TmpRSA);
-		SSL_set_tmp_dh_callback(accepted->transport.tls.ssl,  ssl_callback_TmpDH);
-
-
-		ev_timer_init(&accepted->events.handshake_timeout, connection_tls_handshake_again_timeout_cb, 0., con->events.handshake_timeout.repeat);
-//		ev_timer_init(&accepted->events.idle_timeout, connection_tls_accept_again_timeout_cb, 0., con->events.idle_timeout.repeat);
-
-
-		// create protocol specific data
-		accepted->protocol.ctx = accepted->protocol.ctx_new(accepted);
-
-
-		accepted->stats.io_in.throttle.max_bytes_per_second = con->stats.io_in.throttle.max_bytes_per_second;
-		accepted->stats.io_out.throttle.max_bytes_per_second = con->stats.io_out.throttle.max_bytes_per_second;
-
-		// teach new connection about parent
-		if( con->protocol.origin != NULL )
-			con->protocol.origin(accepted, con);
-
-		connection_set_state(accepted, connection_state_handshake);
-		SSL_set_accept_state(accepted->transport.tls.ssl);
-
-		accepted->events.io_in.events = EV_READ;
-		connection_tls_handshake_again_cb(EV_A_ &accepted->events.io_in, 0);
-
-		i = incident_new("dionaea.connection.link");
-		incident_value_con_set(i, "parent", con);
-		incident_value_con_set(i, "child", accepted);
-		incident_report(i);
-		incident_free(i);
-	}
-
-	if( ev_is_active(&con->events.listen_timeout) )
-	{
-		ev_clear_pending(EV_A_ &con->events.listen_timeout);
-		ev_timer_again(EV_A_  &con->events.listen_timeout);
-	}
-}
-
-
-void connection_tls_handshake_again_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = NULL;
-	struct incident *i;
-
-	if( w->events == EV_READ )
-		con = CONOFF_IO_IN(w);
-	else
-		con	= CONOFF_IO_OUT(w);
-	g_debug("%s con %p %i %p %p",__PRETTY_FUNCTION__, con, revents, CONOFF_IO_IN(w), CONOFF_IO_OUT(w));
-
-	ev_io_stop(EV_A_ &con->events.io_in);
-	ev_io_stop(EV_A_ &con->events.io_out);
-
-	int err = SSL_do_handshake(con->transport.tls.ssl);
-	connection_tls_error(con);
-	if( err != 1 )
-	{
-		g_debug("setting connection_tls_accept_again_timeout_cb to %f",con->events.handshake_timeout.repeat);
-		ev_timer_again(EV_A_ &con->events.handshake_timeout);
-
-		int action = SSL_get_error(con->transport.tls.ssl, err);
-		g_debug("SSL_do_handshake failed %i %i read:%i write:%i", err, action, SSL_ERROR_WANT_READ, SSL_ERROR_WANT_WRITE);
-
-		connection_tls_error(con);
-		switch( action )
-		{
-		case SSL_ERROR_NONE:
-			g_debug("SSL_ERROR_NONE %s:%i", __FILE__,  __LINE__);
-			break;
-		case SSL_ERROR_ZERO_RETURN:
-			g_debug("SSL_ERROR_ZERO_RETURN %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-			ev_io_init(&con->events.io_in, connection_tls_handshake_again_cb, con->socket, EV_READ);
-			ev_io_start(EV_A_ &con->events.io_in);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			ev_io_init(&con->events.io_out, connection_tls_handshake_again_cb, con->socket, EV_WRITE);
-			ev_io_start(EV_A_ &con->events.io_out);
-			break;
-
-		case SSL_ERROR_WANT_ACCEPT:
-			g_debug("SSL_ERROR_WANT_ACCEPT%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_X509_LOOKUP:
-			g_debug("SSL_ERROR_WANT_X509_LOOKUP %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SYSCALL:
-			g_debug("SSL_ERROR_SYSCALL %s:%i", __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-
-		case SSL_ERROR_SSL:
-			g_debug("SSL_ERROR_SSL %s:%i", __FILE__,  __LINE__);
-			connection_tls_disconnect(con);
-			break;
-		}
-	} else
-	{
-		g_debug("SSL_do_handshake success");
-		ev_timer_stop(EV_A_ &con->events.handshake_timeout);
-		ev_timer_init(&con->events.idle_timeout, connection_idle_timeout_cb, 0. ,con->events.idle_timeout.repeat);
-		connection_established(con);
-
-		i = incident_new("dionaea.connection.tls.accept");
-		incident_value_con_set(i, "con", con);
-		incident_report(i);
-		incident_free(i);
-	}
-}
-
-void connection_tls_handshake_again_timeout_cb(EV_P_ struct ev_timer *w, int revents)
-{
-	struct connection *con = CONOFF_HANDSHAKE_TIMEOUT(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	switch( con->type )
-	{
-	case connection_type_connect:
-		ev_timer_stop(EV_A_ &con->events.handshake_timeout);
-		ev_io_stop(EV_A_ &con->events.io_out);
-		close(con->socket);
-		con->socket = -1;
-		connection_connect_next_addr(con);
-		break;
-	case connection_type_accept:
-		connection_tls_disconnect(con);
-		break;
-	case connection_type_listen:
-	case connection_type_bind:
-	case connection_type_none:
-		break;
-	}
-}
-
-void connection_tls_disconnect(struct connection *con)
-{
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	enum connection_state state = con->state;
-	connection_set_state(con, connection_state_close);
-
-	connection_disconnect(con);
-
-	g_string_erase(con->transport.tls.io_in, 0, -1);
-	g_string_erase(con->transport.tls.io_out, 0, -1);
-	g_string_erase(con->transport.tls.io_out_again, 0, -1);
-	con->transport.tls.io_out_again_size = 0;
-
-
-	if( con->protocol.disconnect != NULL &&
-		(state != connection_state_none &&
-		 state != connection_state_connecting &&
-		 state != connection_state_handshake) )
-	{
-		bool reconnect = con->protocol.disconnect(con, con->protocol.ctx);
-		g_debug("reconnect is %i", reconnect);
-		if( reconnect == true && con->type == connection_type_connect )
-		{
-			connection_reconnect(con);
-			return;
-		}
-	}
-	connection_free(con);
-}
-
-void connection_tls_connecting_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_OUT(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	ev_timer_stop(EV_A_ &con->events.connecting_timeout);
-
-	int socket_error = 0;
-	int error_size = sizeof(socket_error);
-
-
-	int ret = getsockopt(con->socket, SOL_SOCKET, SO_ERROR, &socket_error,(socklen_t *)&error_size);
-
-	if( ret != 0 || socket_error != 0 )
-	{
-		errno = socket_error;
-		ev_io_stop(EV_A_ &con->events.io_out);
-		close(con->socket);
-		connection_connect_next_addr(con);
-		return;
-	}
-
-	connection_node_set_local(con);
-	connection_node_set_remote(con);
-
-	g_debug("connection %s -> %s", con->local.node_string, con->remote.node_string);
-
-	if( con->transport.tls.ssl != NULL )
-		SSL_free(con->transport.tls.ssl);
-
-	con->transport.tls.ssl = SSL_new(con->transport.tls.ctx);
-	SSL_set_fd(con->transport.tls.ssl, con->socket);
-
-	ev_timer_init(&con->events.handshake_timeout, connection_tls_handshake_again_timeout_cb, 0., con->events.handshake_timeout.repeat);
-
-	connection_set_state(con, connection_state_handshake);
-
-	SSL_set_connect_state(con->transport.tls.ssl);
-
-	con->events.io_in.events = EV_READ;
-	connection_tls_handshake_again_cb(EV_A_ &con->events.io_in, 0);
-}
-
-void connection_tls_error(struct connection *con)
-{
-	con->transport.tls.ssl_error = ERR_get_error();
-	ERR_error_string(con->transport.tls.ssl_error, con->transport.tls.ssl_error_string);
-	if( con->transport.tls.ssl_error != 0 )
-		g_debug("SSL ERROR %s\t%s", con->transport.tls.ssl_error_string, SSL_state_string_long(con->transport.tls.ssl));
-}
-
-/*
- *
- * connection udp
- *
- */
-
-
-ssize_t recvfromto(int sockfd, void *buf, size_t len, int flags,
-				 const struct sockaddr *fromaddr, socklen_t *fromlen,
-				 const struct sockaddr *toaddr, socklen_t *tolen)
-{
-	struct iovec iov[1];
-#if defined(IPV6_PKTINFO)
-	char cmsg[CMSG_SPACE(sizeof(struct in6_pktinfo))];
-#else
-	char cmsg[CMSG_SPACE(64)];
-#endif
-	struct cmsghdr *cmsgptr;
-	struct msghdr msg;
-	ssize_t rlen;
-
-	iov[0].iov_base = buf;
-	iov[0].iov_len = len;
-
-	memset(&msg, 0, sizeof(msg));
-	msg.msg_name = (void *)fromaddr;
-	msg.msg_namelen = *fromlen;
-	msg.msg_iov = iov;
-	msg.msg_iovlen = 1;
-	msg.msg_control = &cmsg;
-	msg.msg_controllen = sizeof(cmsg);
-
-	if( (rlen = recvmsg(sockfd, &msg, 0)) == -1 )
-		return -1;
-
-
-	for( cmsgptr = CMSG_FIRSTHDR(&msg); cmsgptr != NULL; cmsgptr = CMSG_NXTHDR(&msg, cmsgptr) )
-	{
-#ifdef SOL_IP
-		if( fromaddr->sa_family == PF_INET && cmsgptr->cmsg_level == SOL_IP && cmsgptr->cmsg_type == IP_PKTINFO )
-		{ /* IPv4 */
-			if( *fromlen < sizeof(struct sockaddr_in) )
-			{
-				errno = EINVAL;
-				return -1;
-			}
-			void *addr = ADDROFFSET(toaddr);
-			void *t = &((struct in_pktinfo *)(CMSG_DATA(cmsgptr)))->ipi_addr;
-			memcpy(addr, t, sizeof(struct in_addr));
-			break;
-		}else
-#endif
-#if defined(SOL_IPV6) && defined(IPV6_PKTINFO)
-		if( fromaddr->sa_family == PF_INET6 && cmsgptr->cmsg_level == SOL_IPV6 && cmsgptr->cmsg_type == IPV6_PKTINFO )
-		{ /* IPv6 */
-			if( *fromlen < sizeof(struct sockaddr_in6) )
-			{
-				errno = EINVAL;
-				return -1;
-			}
-			void *addr = ADDROFFSET(toaddr);
-			void *t = &((struct in6_pktinfo *)(CMSG_DATA(cmsgptr)))->ipi6_addr;
-			memcpy(addr, t, sizeof(struct in6_addr));
-			break;
-		}
-#endif
-	}
-	return rlen;
-}
-
-ssize_t sendtofrom(int fd, void *buf, size_t len, int flags, struct sockaddr *to, socklen_t tolen, struct sockaddr *from, socklen_t fromlen)
-{
-	struct iovec iov[1];
-	struct msghdr msg;
-
-	struct cmsghdr* cmsgptr;
-	cmsgptr = NULL;
-
-	iov[0].iov_base = buf;
-	iov[0].iov_len = len;
-
-	memset(&msg, 0, sizeof(msg));
-	msg.msg_name = (void *) to;
-	msg.msg_namelen = tolen;
-	msg.msg_iov = iov;
-	msg.msg_iovlen = 1;
-	msg.msg_flags = flags;
-
-
-
-	if( from->sa_family == PF_INET )
-	{ /* IPv4 */
-#if defined(SOL_IP) && defined(IP_PKTINFO)
-		char cbuf[CMSG_SPACE(sizeof(struct in_pktinfo))];
-		memset(cbuf, 0, sizeof(cbuf));
-		msg.msg_control = cbuf;
-		msg.msg_controllen = sizeof(cbuf);
-
-		cmsgptr = CMSG_FIRSTHDR(&msg);
-		cmsgptr->cmsg_level = SOL_IP;
-		cmsgptr->cmsg_type = IP_PKTINFO;
-		cmsgptr->cmsg_len = CMSG_LEN(sizeof(struct in_pktinfo));
-		memcpy(&((struct in_pktinfo *)(CMSG_DATA(cmsgptr)))->ipi_spec_dst.s_addr, ADDROFFSET(from),  sizeof(struct in_addr) );
-		return sendmsg(fd, &msg, 0);
-#endif
-	}else
-	if( from->sa_family == PF_INET6 )
-	{ /* IPv6 */
-#if defined(SOL_IPV6) && defined(IPV6_PKTINFO)
-		char cbuf[CMSG_SPACE(sizeof(struct in6_pktinfo))];
-		memset(cbuf, 0, sizeof(cbuf));
-		msg.msg_control = cbuf;
-		msg.msg_controllen = sizeof(cbuf);
-
-		cmsgptr = CMSG_FIRSTHDR(&msg);
-		cmsgptr->cmsg_level = SOL_IPV6;
-		cmsgptr->cmsg_type = IPV6_PKTINFO;
-		cmsgptr->cmsg_len = CMSG_LEN(sizeof(struct in6_pktinfo));
-		memcpy(&((struct in6_pktinfo *)(CMSG_DATA(cmsgptr)))->ipi6_addr, ADDROFFSET(from),  sizeof(struct in6_addr) );
-		return sendmsg(fd, &msg, 0);
-#endif
-	}else
-	{
-		errno = EINVAL;
-		return -1;
-	}
-	/* if your operating system lacks everything ... */
-	g_warning("Your operating system lacks SOL_IP(V6) / IP(V6)_PKTINFO");
-	return sendto(fd, buf, len, flags, to, tolen);
-}
-
-void connection_udp_io_in_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	socklen_t sizeof_sa = sizeof(struct sockaddr_storage);
-	socklen_t sizeof_sb = sizeof(struct sockaddr_storage);
-	unsigned char buf[64*1024];
-	memset(buf, 0, 64*1024);
-	int ret;
-	while( (ret = recvfromto(con->socket, buf, 64*1024, 0,  (struct sockaddr *)&con->remote.addr, &sizeof_sa, (struct sockaddr *)&con->local.addr, &sizeof_sb)) > 0 )
-	{
-		node_info_set(&con->remote, &con->remote.addr);
-		node_info_set(&con->local, &con->local.addr);
-
-
-		struct connection *peer;
-		switch( con->type )
-		{
-		case connection_type_listen:
-			/* we are a server -> find the peer */
-			if( (peer = g_hash_table_lookup(con->transport.udp.type.server.peers, con)) == NULL )
-			{ /* no peer? create a new one */
-				peer = connection_new(connection_transport_udp);
-				peer->transport.udp.type.client.parent = con;
-				peer->type = connection_type_accept;
-				peer->state = connection_state_established;
-				memcpy(&peer->local.addr, &con->local.addr, sizeof(struct sockaddr_storage));
-				memcpy(&peer->remote.addr, &con->remote.addr, sizeof(struct sockaddr_storage));
-				node_info_set(&peer->remote, &peer->remote.addr);
-				node_info_set(&peer->local, &peer->local.addr);
-				g_debug("new udp peer %s %s", peer->local.node_string, peer->remote.node_string);
-				peer->transport.udp.type.client.parent = con;
-				connection_protocol_set(peer, &peer->transport.udp.type.client.parent->protocol);
-
-				ev_timer_init(&peer->events.idle_timeout, connection_idle_timeout_cb, 0. ,con->events.idle_timeout.repeat);
-				ev_timer_init(&peer->events.sustain_timeout, connection_sustain_timeout_cb, 0. ,con->events.sustain_timeout.repeat);
-				peer->protocol.ctx = peer->transport.udp.type.client.parent->protocol.ctx_new(peer);
-
-				// teach new connection about parent
-				if( peer->transport.udp.type.client.parent->protocol.origin != NULL )
-					peer->transport.udp.type.client.parent->protocol.origin(peer, peer->transport.udp.type.client.parent);
-
-				connection_established(peer);
-
-				g_hash_table_insert(con->transport.udp.type.server.peers, peer, peer);
-			}
-			g_debug("%s -> %s %i bytes", peer->remote.node_string, peer->local.node_string, ret);
-			break;
-		case connection_type_bind:
-		case connection_type_connect:
-			peer = con;
-			break;
-		default:
-			return;
-		}
-
-		if( peer->type == connection_type_accept && peer->processor_data != NULL )
-			processors_io_in(peer, buf, ret);
-
-		peer->protocol.io_in(peer, peer->protocol.ctx, buf, ret);
-	}
-
-	if( ret == -1 && errno != EAGAIN && errno != EWOULDBLOCK )
-	{
-		g_warning("connection error %i %s", ret, strerror(errno));
-	}
-}
-
-void _connection_send_packets(struct connection *con, int fd, GList **packets)
-{
-	GList *elem;
-
-	while( (elem = g_list_first(*packets)) != NULL )
-	{
-		struct udp_packet *packet = elem->data;
-		socklen_t size = ((struct sockaddr *)&packet->to)->sa_family == PF_INET ? sizeof(struct sockaddr_in) :
-						 ((struct sockaddr *)&packet->to)->sa_family == PF_INET6 ? sizeof(struct sockaddr_in6) :
-						 ((struct sockaddr *)&packet->to)->sa_family == AF_UNIX ? sizeof(struct sockaddr_un) : -1;
-
-		int ret;
-		/*
-		 * for whatever reason
-		 * * send
-		 *   - works on udp sockets which were connect()'ed before (linux)
-		 *   - works not on udp sockets which were bind()'ed before (linux)
-		 * * sendto
-		 *   - does not work on connect()'ed sockets on (openbsd)
-		 *
-		 * and as we can't distinguish from bound/unbound connected/unconnected sockets at this point
-		 * udp does not work for openbsd
-		 */
-		if( con->type == connection_type_accept && con->processor_data != NULL )
-			processors_io_out(con, packet->data->str, packet->data->len);
-
-		ret = sendtofrom(fd, packet->data->str, packet->data->len, 0, (struct sockaddr *)&packet->to, size, (struct sockaddr *)&packet->from, size);
-
-		if( ret == -1 )
-		{
-			if( errno == EAGAIN )
-			{
-				break;
-			} else
-			{
-				g_debug("domain %i size %i", ((struct sockaddr *)&packet->to)->sa_family, size);
-				g_warning("sendtofrom failed %s",  strerror(errno));
-				g_string_free(packet->data, TRUE);
-				g_free(packet);
-				*packets = g_list_delete_link(*packets, elem);
-			}
-			break;
-		} else
-		if( ret == packet->data->len )
-		{
-			g_string_free(packet->data, TRUE);
-			g_free(packet);
-			*packets = g_list_delete_link(*packets, elem);
-		} else
-		{
-			g_warning("sendtofrom failed %s",  strerror(errno));
-			g_string_free(packet->data, TRUE);
-			g_free(packet);
-			*packets = g_list_delete_link(*packets, elem);
-			break;
-		}
-	}
-}
-
-void connection_udp_io_out_cb(EV_P_ struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_OUT(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	int fd = 0;
-	switch( con->type )
-	{
-	case connection_type_connect:
-	case connection_type_bind:
-		fd = con->socket;
-		break;
-	case connection_type_accept:
-		fd = con->transport.udp.type.client.parent->socket;
-		break;
-	default:
-		g_warning("Invalid connection type!");
-	}
-
-	_connection_send_packets(con, fd, &con->transport.udp.io_out);
-
-	if( g_list_length(con->transport.udp.io_out) > 0 )
-	{
-		if( !ev_is_active(&con->events.io_out) )
-		{
-			ev_io_init(&con->events.io_out, connection_udp_io_out_cb, con->socket, EV_WRITE);
-			ev_io_start(EV_A_ &con->events.io_out);
-		}
-	} else
-	{
-		ev_io_stop(EV_A_ &con->events.io_out);
-	}
-	if( ev_is_active(&con->events.idle_timeout) )
-		ev_timer_again(CL, &con->events.idle_timeout);
-}
-
-void connection_udp_disconnect(struct connection *con)
-{
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	connection_set_state(con, connection_state_close);
-	switch( con->type )
-	{
-	case connection_type_accept:
-		g_hash_table_remove(con->transport.udp.type.client.parent->transport.udp.type.server.peers, con);
-	case connection_type_connect:
-	case connection_type_bind:
-		con->protocol.disconnect(con, con->protocol.ctx);
-		break;
-	case connection_type_listen:
-		{
-			GHashTableIter iter;
-			gpointer key, value;
-			g_hash_table_iter_init (&iter, con->transport.udp.type.server.peers);
-			while( g_hash_table_iter_next (&iter, &key, &value) )
-			{
-				struct connection *peer = value;
-				connection_udp_disconnect(peer);
-				g_hash_table_iter_init (&iter, con->transport.udp.type.server.peers);
-			}
-		}
-		break;
-	default:
-		break;
-	}
-
-	connection_disconnect(con);
-	connection_free(con);
-}
-
-/*
- * dtls
- */
-
-int _SSL_connection_index;
-
-bool connection_dtls_mkcert(struct connection *con)
-{
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	return mkcert(con->transport.dtls.ctx);
-}
-
-guint connection_addrs_hash(gconstpointer key)
-{
-	g_debug("%s con %p", __PRETTY_FUNCTION__, key);
-	const struct connection *con = key;
-	int local = con->local.port;
-	int remote = (con->remote.port << 16);
-	return local | remote;
-}
-
-gboolean connection_addrs_cmp(gconstpointer a, gconstpointer b)
-{
-	const struct connection *ca = a;
-	const struct connection *cb = b;
-	g_debug("%s con %p %p %s %s | %s %s", __PRETTY_FUNCTION__, a, b,
-			ca->local.node_string, cb->local.node_string,
-			ca->remote.node_string, cb->remote.node_string);
-	if( memcmp(&ca->local.addr, &cb->local.addr, sizeof(struct sockaddr_storage)) == 0 &&
-		memcmp(&ca->remote.addr, &cb->remote.addr, sizeof(struct sockaddr_storage)) == 0)
-		return true;
-	return false;
-}
-
-void dtls_create_cookie(struct connection *con, unsigned char *hash, unsigned int *len)
-{
-	/* Create buffer with peer's address and port */
-	g_debug("%s con %p %s:%s", __PRETTY_FUNCTION__, con, con->remote.ip_string, con->remote.port_string);
-	int length = sizeof(in_port_t) + ADDRSIZE(&con->remote.addr);
-	unsigned char buffer[length];
-	memcpy(buffer, PORTOFFSET(&con->remote.addr), sizeof(in_port_t));
-	memcpy(buffer + sizeof(in_port_t), ADDROFFSET(&con->remote.addr), ADDRSIZE(&con->remote.addr));
-
-	/* Calculate HMAC of buffer using the secret */
-	HMAC(EVP_sha1(),
-		 (const void*) con->transport.dtls.type.client.parent->transport.dtls.type.server.cookie_secret, DTLS_COOKIE_SECRET_LENGTH,
-		 (const unsigned char*) buffer, length,
-		 hash, len);
-}
-
-/**
- *
- *
- * @param ssl
- * @param cookie
- * @param cookie_len
- *
- * @return int 0 on error
- */
-int dtls_generate_cookie_cb(SSL *ssl, unsigned char *cookie, unsigned int *len)
-{
-	struct connection *con = SSL_get_ex_data(ssl, _SSL_connection_index);
-	g_debug("%s con %p", __PRETTY_FUNCTION__, con);
-	dtls_create_cookie(con, cookie, len);
-	return 1;
-}
-
-/**
- *
- *
- * @param ssl
- * @param cookie
- * @param cookie_len
- *
- * @return int 0 on error
- */
-int dtls_verify_cookie_cb(SSL *ssl, unsigned char *cookie, unsigned int len)
-{
-	struct connection *con = SSL_get_ex_data(ssl, _SSL_connection_index);
-	g_debug("%s con %p cookie %p len %i", __PRETTY_FUNCTION__, con, cookie, len);
-
-	unsigned char _cookie[SHA_DIGEST_LENGTH];
-	unsigned int _len;
-	con->transport.dtls.type.client.flags |= DTLS_HAS_SEEN_THE_COOKIE;
-	dtls_create_cookie(con, _cookie, &_len);
-
-	if( len == _len && memcmp(_cookie, cookie, len) == 0 )
-		return 1;
-	return 0;
-}
-
-void connection_dtls_error(struct connection *con)
-{
-	con->transport.dtls.ssl_error = ERR_get_error();
-	ERR_error_string(con->transport.dtls.ssl_error, con->transport.dtls.ssl_error_string);
-	if( con->transport.dtls.ssl_error != 0 )
-		g_debug("SSL ERROR %s\t%s", con->transport.dtls.ssl_error_string, SSL_state_string_long(con->transport.dtls.ssl));
-}
-
-void connection_dtls_connect_again(struct ev_loop *loop, struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	g_debug("CLIENT state %s", SSL_state_string_long(con->transport.dtls.ssl));
-	int err = SSL_connect(con->transport.dtls.ssl);
-	connection_dtls_drain_bio(con);
-
-	if( err != 1 )
-		return;
-	g_warning("connected!");
-	connection_established(con);
-	ev_io_start(EV_A_  &con->events.io_in);
-}
-
-void connection_dtls_accept_again(struct ev_loop *loop, struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	g_debug("CLIENT state %s", SSL_state_string_long(con->transport.dtls.ssl));
-	int err = SSL_accept(con->transport.dtls.ssl);
-
-	if( err != 1 )
-	{
-		int action = SSL_get_error(con->transport.dtls.ssl, err);
-		g_debug("SSL_accept failed %i %i read:%i write:%i", err, action, SSL_ERROR_WANT_READ, SSL_ERROR_WANT_WRITE);
-		connection_dtls_error(con);
-		if( !(con->transport.dtls.type.client.flags & DTLS_HAS_SEEN_THE_COOKIE) )
-		{
-			g_warning("CLIENT CONNECT WITHOUT COOKIE!");
-			g_hash_table_remove(con->transport.dtls.type.client.parent->transport.dtls.type.server.peers, con);
-			connection_dtls_drain_bio(con);
-			connection_free_cb(EV_A_ &con->events.free, 0, true);
-			return;
-		}
-		switch( action )
-		{
-		case SSL_ERROR_NONE:
-			g_debug("%s:%i", __FILE__,  __LINE__);
-			break;
-		case SSL_ERROR_ZERO_RETURN:
-			g_debug("%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-			connection_dtls_drain_bio(con);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			connection_dtls_drain_bio(con);
-			break;
-
-		case SSL_ERROR_WANT_ACCEPT:
-			g_debug("SSL_ERROR_WANT_ACCEPT%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_X509_LOOKUP:
-			g_debug("SSL_ERROR_WANT_X509_LOOKUP %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SYSCALL:
-			g_debug("SSL_ERROR_SYSCALL %s:%i", __FILE__,  __LINE__);
-//			connection_dtls_disconnect(con);
-			break;
-
-		case SSL_ERROR_SSL:
-			g_debug("SSL_ERROR_SSL %s:%i", __FILE__,  __LINE__);
-//			connection_dtls_disconnect(con);
-			break;
-
-		}
-	} else
-	{
-		g_debug("SSL_accept success");
-		ev_timer_stop(EV_A_ &con->events.handshake_timeout);
-
-		// set protocol for accepted connection
-		connection_protocol_set(con, &con->transport.dtls.type.client.parent->protocol);
-
-		// copy connect timeout to new connection
-//		ev_timer_init(&con->events.idle_timeout, connection_dtls_idle_timeout_cb, 0. ,con->events.idle_timeout.repeat);
-
-		// create protocol specific data
-		con->protocol.ctx = con->transport.dtls.type.client.parent->protocol.ctx_new(con);
-
-		// teach new connection about parent
-		if( con->transport.dtls.type.client.parent->protocol.origin != NULL )
-			con->transport.dtls.type.client.parent->protocol.origin(con, con->transport.dtls.type.client.parent);
-
-		connection_established(con);
-/*
-		struct incident *i;
-		i = incident_new("dionaea.connection.dtls.accept");
-		incident_value_con_set(i, "con", con);
-		incident_report(i);
-		incident_free(i);
-
-		i = incident_new("dionaea.connection.link");
-		incident_value_con_set(i, "parent", con->transport.dtls.type.client.parent);
-		incident_value_con_set(i, "child", con);
-		incident_report(i);
-		incident_free(i);
-*/
-	}
-}
-
-void connection_dtls_io_in(struct ev_loop *loop, struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-	unsigned char buf[64*1024];
-	int r = SSL_read(con->transport.dtls.ssl, buf, 1024*64);
-	if( r > 0)
-	{
-		g_debug("recv %.*s", r, buf);
-		con->protocol.io_in(con, con->protocol.ctx, buf, r);
-	}else
-	{
-		int action = SSL_get_error(con->transport.dtls.ssl, r);
-		connection_dtls_error(con);
-		switch( action )
-		{
-		case SSL_ERROR_NONE:
-			g_debug("SSL_ERROR_NONE %s:%i", __FILE__,  __LINE__);
-			break;
-		case SSL_ERROR_ZERO_RETURN:
-			g_debug("SSL_ERROR_ZERO_RETURN %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_READ:
-			g_debug("SSL_WANT_READ %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_WRITE:
-			g_debug("SSL_WANT_WRITE %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_ACCEPT:
-			g_debug("SSL_ERROR_WANT_ACCEPT%s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_WANT_X509_LOOKUP:
-			g_debug("SSL_ERROR_WANT_X509_LOOKUP %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SYSCALL:
-			g_debug("SSL_ERROR_SYSCALL %s:%i", __FILE__,  __LINE__);
-			break;
-
-		case SSL_ERROR_SSL:
-			g_debug("SSL_ERROR_SSL %s:%i", __FILE__,  __LINE__);
-			break;
-		default:
-			g_debug("SSL_ERROR %i %i",r, action);
-		}
-	}
-}
-
-void connection_dtls_drain_bio(struct connection *con)
-{
-	if( BIO_ctrl_pending(con->transport.dtls.writing) > 0 )
-	{
-		g_warning("need to flush the bio");
-		unsigned char buf[64*1024];
-		uint32_t size = BIO_read(con->transport.dtls.writing, buf, sizeof(buf));
-		{
-			struct udp_packet *packet = g_malloc0(sizeof(struct udp_packet));
-			packet->data = g_string_new_len((void *)buf, size);
-			memcpy(&packet->to, &con->remote.addr, sizeof(struct sockaddr_storage));
-			memcpy(&packet->from, &con->local.addr, sizeof(struct sockaddr_storage));
-			con->transport.dtls.io_out = g_list_append(con->transport.dtls.io_out, packet);
-			connection_dtls_io_out_cb(g_dionaea->loop, &con->events.io_out, 0);
-		}
-	}
-}
-
-void connection_dtls_io_in_cb(struct ev_loop *loop, struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_IN(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	socklen_t sizeof_sa = sizeof(struct sockaddr_storage);
-	socklen_t sizeof_sb = sizeof(struct sockaddr_storage);
-	unsigned char buf[64*1024];
-	memset(buf, 0, 64*1024);
-
-	int ret;
-	while( (ret = recvfromto(con->socket, buf, 64*1024, 0,  (struct sockaddr *)&con->remote.addr, &sizeof_sa, (struct sockaddr *)&con->local.addr, &sizeof_sb)) > 0 )
-	{
-		node_info_set(&con->remote, &con->remote.addr);
-		node_info_set(&con->local, &con->local.addr);
-
-		struct connection *peer;
-		if( con->type == connection_type_listen )
-		{ /* we are a server -> find the peer */
-			if( (peer = g_hash_table_lookup(con->transport.dtls.type.server.peers, con)) == NULL )
-			{ /* no peer? create a new one */
-				peer = connection_new(connection_transport_dtls);
-				peer->transport.dtls.type.client.parent = con;
-				peer->type = connection_type_accept;
-				peer->state = connection_state_handshake;
-				memcpy(&peer->local.addr, &con->local.addr, sizeof(struct sockaddr_storage));
-				memcpy(&peer->remote.addr, &con->remote.addr, sizeof(struct sockaddr_storage));
-				node_info_set(&peer->remote, &peer->remote.addr);
-				node_info_set(&peer->local, &peer->local.addr);
-				g_debug("new dtls peer %s %s", peer->local.node_string, peer->remote.node_string);
-				peer->transport.dtls.type.client.parent = con;
-				peer->transport.dtls.ctx = con->transport.dtls.ctx;
-				peer->transport.dtls.ssl = SSL_new(peer->transport.dtls.ctx);
-#if OPENSSL_VERSION_NUMBER < 0x10100000L // OpenSSL 1.1.0
-				peer->transport.dtls.ssl->d1->listen = 1;
-#endif
-				SSL_CTX_set_session_cache_mode(peer->transport.dtls.ctx, SSL_SESS_CACHE_OFF);
-				peer->transport.dtls.reading = BIO_new(BIO_s_mem());
-				peer->transport.dtls.writing = BIO_new(BIO_s_mem());
-				BIO_set_mem_eof_return(peer->transport.dtls.reading, -1);
-				BIO_set_mem_eof_return(peer->transport.dtls.writing, -1);
-				SSL_set_bio(peer->transport.dtls.ssl, peer->transport.dtls.reading, peer->transport.dtls.writing);
-				SSL_set_accept_state(peer->transport.dtls.ssl);
-#if OPENSSL_VERSION_NUMBER > 0x1000004fL // OpenSSL 1.0.0d 8 Feb 2011
-				SSL_set_options(peer->transport.dtls.ssl, SSL_OP_COOKIE_EXCHANGE);
-#endif
-				SSL_set_ex_data(peer->transport.dtls.ssl, _SSL_connection_index, peer);
-				g_hash_table_insert(con->transport.dtls.type.server.peers, peer, peer);
-			}
-			g_debug("%s -> %s %i bytes", peer->remote.node_string, peer->local.node_string, ret);
-		}else
-		if( con->type == connection_type_connect )
-		{
-			peer = con;
-		}else
-		{
-			g_critical("Invalid connection for DTLS");
-		}
-		BIO_write(peer->transport.dtls.reading, buf, ret);
-
-		switch( peer->state )
-		{
-		case connection_state_handshake:
-			if( peer->type == connection_type_accept )
-				connection_dtls_accept_again(EV_A_ &peer->events.io_in, EV_READ);
-			else
-			if( peer->type == connection_type_connect )
-				connection_dtls_connect_again(EV_A_ &peer->events.io_in, EV_READ);
-			break;
-		case connection_state_established:
-			connection_dtls_io_in(EV_A_ &peer->events.io_in, EV_READ);
-			break;
-		default:
-			g_warning("UNHANDLED STATE");
-		}
-	}
-	if( ret == -1 && errno != EAGAIN && errno != EWOULDBLOCK )
-	{
-		g_warning("connection error %i %s", ret, strerror(errno));
-	}
-}
-
-void connection_dtls_io_out_cb(struct ev_loop *loop, struct ev_io *w, int revents)
-{
-	struct connection *con = CONOFF_IO_OUT(w);
-	g_debug("%s con %p",__PRETTY_FUNCTION__, con);
-
-	int fd = 0;
-	switch( con->type )
-	{
-	case connection_type_connect:
-		fd = con->socket;
-		break;
-	case connection_type_accept:
-		fd = con->transport.dtls.type.client.parent->socket;
-		break;
-	default:
-		g_warning("Invalid connection type!");
-	}
-	_connection_send_packets(con, fd, &con->transport.dtls.io_out);
-
-//	g_debug(" done");
-	if( g_list_length(con->transport.dtls.io_out) > 0 )
-	{
-		if( !ev_is_active(&con->events.io_out) )
-		{
-			ev_io_init(&con->events.io_out, connection_udp_io_out_cb, con->socket, EV_WRITE);
-			ev_io_start(EV_A_ &con->events.io_out);
-		}
-	} else
-	{
-		ev_io_stop(EV_A_ &con->events.io_out);
-	}
-	if( ev_is_active(&con->events.idle_timeout) )
-		ev_timer_again(CL, &con->events.idle_timeout);
 }
